Day 26/requirements.txt

fastapi
uvicorn
sqlalchemy
pymysql
pymongo
python-dotenv
pytest
